Edited README

+++ routes/web.php
@@ -20,7 +20,6 @@ use \App\Http\Controllers\InspectionController;
 Route::get('/', function () {
     return view('welcome');
 });
-Route::get('/proto/{inspection}', ProtocolController::class)->name('proto');
 
 //Route::group(['middleware' => ['auth']], function (){
 Route::middleware(['auth'])->group(function () {
@@ -30,7 +29,7 @@ Route::middleware(['auth'])->group(function () {
             'lifts_total' => App\Models\Lift::all()->count(),
             'lift_managers_total' => App\Models\LiftManager::get()->count(),
             'inspections_total' => \App\Models\Inspection::get()->count()
-            ]);
+        ]);
     })->name('dashboard');
 
     Route::resource('lifts', LiftController::class)->middleware('is_admin');
@@ -39,12 +38,12 @@ Route::middleware(['auth'])->group(function () {
         Route::prefix('lift-managers')->group(function () {
             Route::name('lift-managers.')->group(function () {
                 Route::get('/', 'index')->name('index');
-                Route::get('/create', 'create')->name('create');
-                Route::post('/', 'store')->name('store');
+                Route::get('/create', 'create')->middleware('is_admin')->name('create');
+                Route::post('/', 'store')->middleware('is_admin')->name('store');
                 Route::get('/{liftManager}', 'show')->name('show');
-                Route::get('/{liftManager}/edit', 'edit')->name('edit');
-                Route::match(['put', 'patch'], '/{liftManager}', 'update')->name('update');
-                Route::delete('/{liftManager}', 'destroy')->name('destroy');
+                Route::get('/{liftManager}/edit', 'edit')->middleware('is_admin')->name('edit');
+                Route::match(['put', 'patch'], '/{liftManager}', 'update')->middleware('is_admin')->name('update');
+                Route::delete('/{liftManager}', 'destroy')->middleware('is_admin')->name('destroy');
             });
         });
     });
@@ -62,6 +61,7 @@ Route::middleware(['auth'])->group(function () {
             });
         });
     });
+    Route::get('/proto/{inspection}', ProtocolController::class)->middleware('is_admin')->name('proto');
 
 });
